update(backend): refactored & reconnected to ElephantSQL
