+++ .babelrc
@@ -1,3 +1,16 @@
 {
-  "plugins": ["@babel/plugin-proposal-class-properties"]
+  "presets": [
+    [
+      "@babel/preset-env",
+      {
+        "targets": {
+          "node": "current"
+        }
+      }
+    ]
+  ],
+  "plugins": [
+    "@babel/plugin-proposal-class-properties", 
+    "@babel/plugin-proposal-optional-chaining"
+  ]
 }
