+++ .env.default
@@ -1,3 +1,22 @@
 NODE_ENV=
+PORT=
+DB_URL_BACKUP=
+HOST=
 
-DB_URL=
+
+DEV_DATABASE_URL=
+DEV_USERNAME=
+DEV_PASSWORD=
+DEV_DEB_NAME=
+
+
+TEST_DATABASE_URL=
+TEST_USERNAME=
+TEST_PASSWORD=
+TEST_DB_NAME=
+
+
+DATABASES_URL=
+USERNAME=
+PASSWORD=
+DB_NAME=
