+++ .prettierrc
@@ -1,4 +1,7 @@
 {
-  "printWidth": 100,
-  "singleQuote": true
+  "singleQuote": true,
+  "tabWidth": 4,
+  "bracketSpacing": true,
+  "arrowParens": "always"
+
 }
