+++ package.json
@@ -11,7 +11,7 @@
     "build": "webpack",
     "start": "concurrently --kill-others-on-fail \"yarn start:server\" \"yarn start:client\"",
     "start:client": "webpack --watch",
-    "start:server": "nodemon server/server"
+    "start:server": "nodemon server/"
   },
   "repository": {
     "type": "git",
@@ -58,6 +58,7 @@
     "@babel/cli": "^7.1.2",
     "@babel/core": "^7.1.2",
     "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.8.3",
     "@babel/plugin-transform-react-constant-elements": "^7.0.0",
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-react": "^7.0.0",
@@ -79,7 +80,7 @@
     "prettier": "^1.15.1",
     "redux-devtools-extension": "^2.13.5",
     "sass-loader": "^7.1.0",
-    "sequelize-cli": "^5.2.0",
+    "sequelize-cli": "^5.5.1",
     "style-loader": "^0.23.1",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.9"
