+++ server/models/index.js
@@ -1,0 +1,17 @@
+/* eslint-disable prettier/prettier */
+
+const Application           = require('./Application');
+const ApplicationMetrics    = require('./ApplicationMetrics');
+const Contact               = require('./Contact');
+const EventMetrics          = require('./EventMetrics');
+const FollowUpMetrics       = require('./FollowUpMetrics');
+const User                  = require('./User');
+
+module.exports = {
+    Application,
+    ApplicationMetrics,
+    Contact,
+    EventMetrics,
+    FollowUpMetrics,
+    User
+};
